CNDIT-1103 : Rename the AvroRecord builder method in the Util class and move the Avro Record related model classes to a separate folder.
