+++ data-reporting-service/organization-service/src/main/java/gov/cdc/etldatapipeline/organization/model/avro/DataEnvelope.java
@@ -1,16 +1,16 @@
-package gov.cdc.etldatapipeline.person.model.dto.dataprops;
+package gov.cdc.etldatapipeline.organization.model.avro;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class DataField {
-    private String type;
-    private boolean optional = true;
-    private String field;
+public class DataEnvelope<T> {
+    private DataSchema schema;
+    private T payload;
 }
