+++ data-reporting-service/organization-service/src/main/java/gov/cdc/etldatapipeline/organization/model/avro/DataField.java
@@ -1,18 +1,16 @@
-package gov.cdc.etldatapipeline.organization.model.dto.dataprops;
-
+package gov.cdc.etldatapipeline.organization.model.avro;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.util.Set;
-
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class DataSchema {
+public class DataField {
     private String type;
-    private Set<DataField> fields;
+    private boolean optional = true;
+    private String field;
 }
