+++ data-reporting-service/organization-service/src/main/java/gov/cdc/etldatapipeline/organization/model/avro/DataSchema.java
@@ -1,16 +1,18 @@
-package gov.cdc.etldatapipeline.person.model.dto.dataprops;
+package gov.cdc.etldatapipeline.organization.model.avro;
+
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.util.Set;
 
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class DataEnvelope<T> {
-    private DataSchema schema;
-    private T payload;
+public class DataSchema {
+    private String type;
+    private Set<DataField> fields;
 }
