+++ data-reporting-service/person-service/src/main/java/gov/cdc/etldatapipeline/person/model/avro/DataField.java
@@ -1,16 +1,16 @@
-package gov.cdc.etldatapipeline.organization.model.dto.dataprops;
+package gov.cdc.etldatapipeline.person.model.avro;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class DataEnvelope<T> {
-    private DataSchema schema;
-    private T payload;
+public class DataField {
+    private String type;
+    private boolean optional = true;
+    private String field;
 }
