+++ data-reporting-service/person-service/src/main/java/gov/cdc/etldatapipeline/person/model/avro/DataSchema.java
@@ -1,4 +1,4 @@
-package gov.cdc.etldatapipeline.person.model.dto.dataprops;
+package gov.cdc.etldatapipeline.person.model.avro;
 
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
