+++ data-reporting-service/person-service/src/test/java/gov/cdc/etldatapipeline/person/KafkaStreamsServiceTest.java
@@ -3,7 +3,7 @@ package gov.cdc.etldatapipeline.person;
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import gov.cdc.etldatapipeline.person.model.dto.PersonExtendedProps;
-import gov.cdc.etldatapipeline.person.model.dto.dataprops.DataEnvelope;
+import gov.cdc.etldatapipeline.person.model.avro.DataEnvelope;
 import gov.cdc.etldatapipeline.person.model.dto.patient.PatientElasticSearch;
 import gov.cdc.etldatapipeline.person.model.dto.patient.PatientReporting;
 import gov.cdc.etldatapipeline.person.model.dto.patient.PatientSp;
